% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{bookbuildeR.main}
\alias{bookbuildeR.main}
\title{Invoke the bookBuildeR Tool Chain}
\usage{
bookbuildeR.main(sourceFile, format.in = "markdown", destName = "book",
  destDir)
}
\arguments{
\item{sourceFile}{the path to the source file}

\item{format.in}{the input format}

\item{destName}{the base name of the destination file without extension}

\item{destDir}{the destination directory}
}
\value{
a list where the names are the file types and the keys are the paths
  to the generated files: currently, we generate pdf, epub, and html output.
}
\description{
Invoke the complete bookBuildeR tool chain
}
