% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/azw3.R
\name{calibre.azw3}
\alias{calibre.azw3}
\title{Use Calibre to Generate awz3 from epub}
\usage{
calibre.azw3(epubFile, destName = sub(pattern = "\\\\..*", replacement =
  "", x = basename(epubFile)), destDir = dirname(epubFile))
}
\arguments{
\item{epubFile}{the path to the EPUB file (see \link{pandoc.epub})}

\item{destName}{the base name of the destination file without extension}

\item{destDir}{the destination directory}
}
\value{
the canonical path to the destination file
}
\description{
Use Calibre to Generate awz3 from epub
}
