% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logger.R
\name{check.dir}
\alias{check.dir}
\title{Check that the directory identified by the given path exists}
\usage{
check.dir(path)
}
\arguments{
\item{path}{the path to the directory}
}
\value{
the normalized path
}
\description{
Check that the directory identified by the given path exists.
}
