% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logger.R
\name{check.file}
\alias{check.file}
\title{Check that the file identified by the given path exists}
\usage{
check.file(path, nonZeroSize = TRUE)
}
\arguments{
\item{path}{the path to the file}

\item{nonZeroSize}{enforce that the file size is not zero}
}
\value{
the normalized path
}
\description{
Check that the file identified by the given path exists.
}
