% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codeLoad.R
\name{code.load}
\alias{code.load}
\title{Read Code from a File according to Parameters given as Strings}
\usage{
code.load(path, lines = "", tags = "", basePath = NULL)
}
\arguments{
\item{path}{the path to the file to read}

\item{lines}{a set of selected lines, given as R integer vector expression string}

\item{tags}{the list of marking start and end given as string with comma-separated values}
}
\description{
Read snippets of code from a file. Similar to \code{code.read}, but interpret the parameters from strings.
}
