% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codeRead.R
\name{code.read}
\alias{code.read}
\title{Read Code from a File}
\usage{
code.read(path, lines = NULL, tags = NULL)
}
\arguments{
\item{path}{the path to the file to read}

\item{lines}{a set of selected lines}

\item{tags}{the tags marking start and end}
}
\description{
Read snippets of code from a file.
}
