% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pandoc.R
\name{invoke.pandoc}
\alias{invoke.pandoc}
\title{Invoke Pandoc}
\usage{
invoke.pandoc(sourceFile, destFile, ...)
}
\arguments{
\item{sourceFile}{the path to the source file}

\item{destFile}{the path to the destination file}

\item{...}{the arguments to be passed to pandoc}
}
\value{
the canonical path to the destination file
}
\description{
Invoke Pandoc with the provided parameters.
}
