% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logger.R
\name{logger}
\alias{logger}
\title{A logger function writing a log message to the output}
\usage{
logger(...)
}
\arguments{
\item{...}{the arguments to be \code{cat}'ed to the ouput}
}
\description{
writs a log message with a date and time in front to the outpuut
}
