% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meta.R
\name{meta.commit}
\alias{meta.commit}
\title{Get the Commit Identifier from the Environment}
\usage{
meta.commit()
}
\value{
a string identifying the commit of the repository for which we build
}
\description{
This function will only work if we build with travis.ci. It
  loads the commit identifier from the environment.
}
