% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meta.R
\name{meta.repository}
\alias{meta.repository}
\title{Get the Repository Identifier from the Environment}
\usage{
meta.repository()
}
\value{
a string identifying the repository for which we build
}
\description{
This function will only work if we build with travis.ci. It
  loads the repository identifier from the environment.
}
