% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getmeta.R
\name{metadata.get}
\alias{metadata.get}
\title{Extract YAML Metadata from a Text}
\usage{
metadata.get(text)
}
\arguments{
\item{text}{the text to read the metadata from}
}
\value{
a list with key-value pairs from the YAML metadata
}
\description{
A method to read the key-value pairs from YAML
  metadata in a text.
}
