% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getmeta.R
\name{metadata.getCodeRepo}
\alias{metadata.getCodeRepo}
\title{Get the source code repository if any is specified in the metadata}
\usage{
metadata.getCodeRepo(metadata = NULL)
}
\arguments{
\item{metadata}{the meta data}
}
\value{
the url of the source code repository, \code{NULL} if none is specified
}
\description{
Get the source code repository if any is specified in the metadata
}
