% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readmeta.R
\name{metadata.hasBibliography}
\alias{metadata.hasBibliography}
\title{Check if a bibliography is specified in the metadata}
\usage{
metadata.hasBibliography(metadata = NULL)
}
\arguments{
\item{metadata}{the meta data}
}
\value{
\code{TRUE} if \code{metadata} specifies a bibliography, \code{FALSE}
  otherwise
}
\description{
Check if a bibliography is specified in the metadata
}
