% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readmeta.R
\name{metadata.read}
\alias{metadata.read}
\title{A very Primitive Method to Read YAML Metadata Key-Value Pairs}
\usage{
metadata.read(srcfile)
}
\arguments{
\item{srcfile}{the file to read from}
}
\value{
a list with key-value pairs from the YAML metadata
}
\description{
A very primitive method to read the key-value pairs from YAML
  metadata in a file.
}
