% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epub.R
\name{pandoc.epub}
\alias{pandoc.epub}
\title{Invoke Pandoc to Produce an Electronic Book}
\usage{
pandoc.epub(
  sourceFile,
  destName = sub(pattern = "\\\\..*", replacement = "", x = basename(sourceFile)),
  destDir = dirname(sourceFile),
  format.in = "markdown",
  standalone = TRUE,
  tabstops = 2L,
  toc.print = TRUE,
  toc.depth = 3L,
  crossref = TRUE,
  bibliography = TRUE,
  numberSections = TRUE,
  mathToGraphic = FALSE,
  metadata = NULL
)
}
\arguments{
\item{sourceFile}{the path to the source file}

\item{destName}{the base name of the destination file without extension}

\item{destDir}{the destination directory}

\item{format.in}{the input format}

\item{standalone}{should we produce a stand-alone document?}

\item{tabstops}{the number of spaces with which we replace a tab character,
or \code{NA} to not replace}

\item{toc.print}{print a table of contents?}

\item{toc.depth}{the depth of the table of content to print}

\item{crossref}{use pandoc-crossref?}

\item{bibliography}{do we have a bibliography?}

\item{numberSections}{should sections be numbered?}

\item{mathToGraphic}{should math be converted to graphics? If \code{TRUE},
math is rendered to png images, which takes quite some time. If
\code{FALSE}, then math is rendered as MathML, which should, by now, be
supported by many EPUB readers. As of version \code{0.8.3}, we stick with
the latter.}

\item{metadata}{the metadata}
}
\value{
the canonical path to the destination file
}
\description{
Invoke Pandoc with the provided parameters.
}
