% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/html.R
\name{pandoc.html}
\alias{pandoc.html}
\title{Invoke Pandoc to Produce a \code{HTML-5} File}
\usage{
pandoc.html(
  sourceFile,
  destName = sub(pattern = "\\\\..*", replacement = "", x = basename(sourceFile)),
  destDir = dirname(sourceFile),
  format.in = "markdown",
  standalone = TRUE,
  tabstops = 2L,
  toc.print = TRUE,
  toc.depth = 3L,
  crossref = TRUE,
  bibliography = TRUE,
  numberSections = TRUE,
  metadata = NULL
)
}
\arguments{
\item{sourceFile}{the path to the source file}

\item{destName}{the base name of the destination file without extension}

\item{destDir}{the destination directory}

\item{format.in}{the input format}

\item{standalone}{should we produce a stand-alone document?}

\item{tabstops}{the number of spaces with which we replace a tab character,
or \code{NA} to not replace}

\item{toc.print}{print a table of contents?}

\item{toc.depth}{the depth of the table of content to print}

\item{crossref}{use pandoc-crossref?}

\item{bibliography}{do we have a bibliography?}

\item{numberSections}{should sections be numbered?}

\item{metadata}{the metadata}
}
\value{
the canonical path to the destination file
}
\description{
Invoke Pandoc with the provided parameters.
}
