% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pandoc.R
\name{pandoc.invoke}
\alias{pandoc.invoke}
\title{Invoke Pandoc}
\usage{
pandoc.invoke(
  sourceFile,
  destFileName,
  destDir = dirname(sourceFile),
  format.in = "markdown",
  format.out = "latex",
  standalone = TRUE,
  tabstops = 2L,
  toc.print = TRUE,
  toc.depth = 3L,
  crossref = TRUE,
  bibliography = TRUE,
  template = NA_character_,
  numberSections = TRUE,
  ...
)
}
\arguments{
\item{sourceFile}{the path to the source file}

\item{destFileName}{the path to the destination file}

\item{destDir}{the destination directory}

\item{format.in}{the input format}

\item{format.out}{the output format}

\item{standalone}{should we produce a stand-alone document?}

\item{tabstops}{the number of spaces with which we replace a tab character,
or \code{NA} to not replace}

\item{toc.print}{print a table of contents?}

\item{toc.depth}{the depth of the table of content to print}

\item{crossref}{use pandoc-crossref?}

\item{bibliography}{do we have a bibliography?}

\item{template}{the template to be used, or \code{NA} if none}

\item{numberSections}{should sections be numbered?}

\item{...}{the arguments to be passed to pandoc}
}
\value{
the canonical path to the destination file
}
\description{
Invoke Pandoc with the provided parameters.
}
