% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessorCommand.R
\name{preprocess.command}
\alias{preprocess.command}
\title{Create a Command based on a Regular Expression}
\usage{
preprocess.command(prefix = "", n = 1L, func, char.open = "{",
  char.close = "}", stripWhiteSpace = FALSE)
}
\arguments{
\item{prefix}{the prefix of the command}

\item{n}{the number of parameters}

\item{func}{the function to be invoked}

\item{char.open}{the opening character}

\item{char.close}{the closing character}

\item{stripWhiteSpace}{should the white space aroung the command be stripped}
}
\description{
Create a function that can be applied to a string and performs a
  regular expression replacement based on invoking a function \code{func}.
}
