% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessorCommand.R
\name{preprocess.command.regexp}
\alias{preprocess.command.regexp}
\title{Create a Regular Expression for a LaTeX-Style Command}
\usage{
preprocess.command.regexp(prefix = "", n = 1L,
  stripWhiteSpace = FALSE)
}
\arguments{
\item{prefix}{the prefix of the command}

\item{n}{the number of parameters}

\item{stripWhiteSpace}{should the white space aroung the command be stripped}
}
\description{
Create a regular expression that can parse LaTeX commands.
}
