% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessorInput.R
\name{preprocess.input}
\alias{preprocess.input}
\title{Recursively Load a Directory Structure}
\usage{
preprocess.input(sourceFile)
}
\arguments{
\item{sourceFile}{the source file path}
}
\value{
the text string resulting from loading the file
}
\description{
Load a source file and return it as text string, while
  recursively solving commands \code{relative.path} and
  \code{relative.input}.
}
