% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessorPlain.R
\name{preprocess.plain}
\alias{preprocess.plain}
\title{Expand the Simple Commands in via Regular Expressions}
\usage{
preprocess.plain(text)
}
\arguments{
\item{text}{the text to process}
}
\value{
the processed text
}
\description{
Expand simple commands, like \code{meta.time}.
}
