% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessorRegexp.R
\name{preprocess.regexp.groups}
\alias{preprocess.regexp.groups}
\title{Pre-process a text with regular expression that contains at least one group}
\usage{
preprocess.regexp.groups(regex, text, func, ...)
}
\arguments{
\item{regex}{the regular expression}

\item{text}{the text}

\item{func}{the function which receives the strings matched to the groups as
vector}

\item{...}{the parameters to be passed to \code{func}}
}
\value{
the processed string
}
\description{
replace all instances of a regular expression in a text via a
  function
}
