% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessorRepo.R
\name{preprocess.repo}
\alias{preprocess.repo}
\title{Pre-Process the Code Repository}
\usage{
preprocess.repo(text, metadata)
}
\arguments{
\item{text}{the basic text}

\item{metadata}{the parsed YAML meta data}
}
\value{
the preprocessed text
}
\description{
Pre-process the code repository part of the documents. If a code
  repository is specified, download the repository, load the code, delete the
  repository.
}
