% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessorTextblock.R
\name{preprocess.textblocks}
\alias{preprocess.textblocks}
\title{Process all Text Blocks in a Markdown file}
\usage{
preprocess.textblocks(text)
}
\arguments{
\item{text}{the array of text}
}
\value{
the output text
}
\description{
Emulate environments like "proof" or "definition".
}
