% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rexec.R
\name{r.exec}
\alias{r.exec}
\title{Execute the provided R code and return the output}
\usage{
r.exec(code)
}
\arguments{
\item{code}{the R code}
}
\description{
Execute the R code provided as input and return a string
  containing all of its output. If the code prints output, say with
  \code{cat}, e.g., as in \code{"cat(5);cat(6)"}, then that output is
  returned as single collapsed string. If the code just returns a value,
  e.g., \code{"5"}, then this value will be converted to a collapsed string
  and returned.
}
