% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rsource.R
\name{r.source}
\alias{r.source}
\title{Execute the provided R file and return the output}
\usage{
r.source(path, basePath = NULL)
}
\arguments{
\item{path}{the path}

\item{basePath}{the base path}
}
\description{
Execute the R file provided as input and return a string
  containing all of its output. If the code prints output, say with
  \code{cat}, e.g., as in \code{"cat(5);cat(6)"}, then that output is
  returned as single collapsed string. If the code just returns a value,
  e.g., \code{"5"}, then this value will be converted to a collapsed string
  and returned.
  This is similar to what \code{r.exec} is doing, but it obtains its input
  from a file.
}
