% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/templates.R
\name{template.load}
\alias{template.load}
\title{Find a Template, Load it into the Specified Directory if Necessary}
\usage{
template.load(template, dir = getwd())
}
\arguments{
\item{template}{the template}

\item{dir}{the destination directory}
}
\value{
the fully qualified path to the template
}
\description{
Following a heuristic, try to load a template into the given
  directory
}
