% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codeListing.R
\name{code.listing}
\alias{code.listing}
\title{Read Code from a File according to Parameters given as Strings and
 Put it into a Listings Environment}
\usage{
code.listing(label = "", caption = "", language = "", path,
  lines = "", tags = "", basePath = NULL, repo = NULL,
  removeMetaComments = TRUE)
}
\arguments{
\item{label}{the label to use}

\item{caption}{the caption to put}

\item{language}{the programming language}

\item{path}{the path to the file to read}

\item{lines}{a set of selected lines, given as R integer vector expression
string}

\item{tags}{the list of marking start and end given as string with
comma-separated values}

\item{basePath}{the base path against which the \code{path} should be
resolved}

\item{repo}{a link to the repository, if any is provided}
}
\description{
Read snippets of code from a file, put it into a code environment.
If a github code repository is specified, create a source code link.
}
