% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/git.R
\name{git.clone}
\alias{git.clone}
\title{A Command for Cloning a Git Repository into a New Temporary Directory}
\usage{
git.clone(repo)
}
\arguments{
\item{repo}{the repository url}
}
\value{
the path to the cloned repository
}
\description{
Use the \code{git} system command to clone/download a git
  repository.
}
