% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pandoc.R
\name{pandoc.invoke}
\alias{pandoc.invoke}
\title{Invoke Pandoc}
\usage{
pandoc.invoke(sourceFile, destFile, format.in = "markdown",
  format.out = "latex", standalone = TRUE, tabstops = 2L,
  toc.print = TRUE, toc.depth = 3L, crossref = TRUE,
  bibliography = TRUE, ...)
}
\arguments{
\item{sourceFile}{the path to the source file}

\item{destFile}{the path to the destination file}

\item{format.in}{the input format}

\item{format.out}{the output format}

\item{standalone}{should we produce a stand-alone document?}

\item{tabstops}{the number of spaces with which we replace a tab character,
or \code{NA} to not replace}

\item{toc.print}{print a table of contents?}

\item{toc.depth}{the depth of the table of content to print}

\item{crossref}{use pandoc-crossref?}

\item{bibliography}{do we have a bibliography?}

\item{...}{the arguments to be passed to pandoc}
}
\value{
the canonical path to the destination file
}
\description{
Invoke Pandoc with the provided parameters.
}
